# docs/openAPI.yaml
openapi: 3.1.0
info:
  title: Rento API
  version: 0.1.0
  summary: Draft API contract generated from project documentation and UML use cases.
  description: |
    This OpenAPI draft was derived from the documentation in `docs/`.

    Main alignment decisions:
    - `deal` and `booking` flows are unified under the `Booking` resource.
    - authentication is modeled as Bearer token based, although the docs mention both sessions and tokens.
    - favorites are scoped to the authenticated user even where the UML uses `user_id` in query/body.
    - public, authenticated, moderation, and internal endpoints are separated by tags.
    - **MVP (listings):** after `POST /listings/{listingId}/publish` the listing becomes `ACTIVE` immediately (no pre-publish moderation queue). The `PENDING_MODERATION` value and all `/moderation/*` operations are **reserved for a future stage** and are not required for MVP.

    The file is intended as a practical baseline for backend/frontend development and can be tightened later.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /listings:
    get:
      tags: [Listings]
      summary: List listings
      security: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: ownerId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ListingStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Listing collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingCollectionResponse'
    post:
      tags: [Listings]
      summary: Create a new listing draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreateRequest'
      responses:
        '201':
          description: Listing draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingCreateResponse'
        '422':
          $ref: '#/components/responses/BadRequest'
...
  /listings/{listingId}/photos:
    get:
      tags: [Listings]
      summary: List listing photos
      security: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing photos
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListingPhoto'
    post:
      tags: [Listings]
      summary: Upload a listing photo
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                isPrimary:
                  type: boolean
                order:
                  type: integer
      responses:
        '201':
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingPhotoUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /listings/{listingId}/publish:
    post:
      tags: [Listings]
      summary: Publish listing immediately
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing status changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingPublishResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ListingId:
      name: listingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: State conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ListingStatus:
      type: string
      description: |
        MVP: typical owner flow is `DRAFT` → `ACTIVE` after successful publish (with at least one photo).
        `PENDING_MODERATION` is reserved for a future pre-publish or async moderation queue and is not used in the MVP publish path.
      enum: [DRAFT, PENDING_MODERATION, ACTIVE, ARCHIVED, BLOCKED]
    RentalPeriod:
      type: string
      enum: [HOUR, DAY, WEEK, MONTH]
...
    Category:
      type: object
      required: [id, name, slug, isActive]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
          nullable: true
        parentId:
          type: string
          format: uuid
          nullable: true
        order:
          type: integer
        isActive:
          type: boolean
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    ListingPhoto:
      type: object
      required: [id, url, order, isPrimary]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        order:
          type: integer
        isPrimary:
          type: boolean
        uploadedAt:
          type: string
          format: date-time
    ListingSummary:
      type: object
      required: [id, ownerId, categoryId, title, description, rentalPrice, rentalPeriod, depositAmount, status]
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        rentalPrice:
          type: number
          format: double
        rentalPeriod:
          $ref: '#/components/schemas/RentalPeriod'
        depositAmount:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/ListingStatus'
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListingDetail:
      allOf:
        - $ref: '#/components/schemas/ListingSummary'
        - type: object
          properties:
            photos:
              type: array
              items:
                $ref: '#/components/schemas/ListingPhoto'
            category:
              $ref: '#/components/schemas/Category'
...
    ListingCreateResponse:
      allOf:
        - $ref: '#/components/schemas/ListingDetail'
        - type: object
          required: [nextStep, message]
          properties:
            nextStep:
              type: string
              enum: [upload_photos]
            message:
              type: string
    ListingPhotoUploadResponse:
      type: object
      required: [photo, totalPhotos, nextStep, message]
      properties:
        photo:
          $ref: '#/components/schemas/ListingPhoto'
        totalPhotos:
          type: integer
        nextStep:
          type: string
          enum: [publish_listing]
        message:
          type: string
    ListingPublishResponse:
      type: object
      required: [id, status, nextStep, message]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ListingStatus'
        nextStep:
          type: string
          nullable: true
        message:
          type: string
...
    ListingCollectionResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ListingSummary'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    ListingCreateRequest:
      type: object
      required: [categoryId, title, description, rentalPrice, rentalPeriod, depositAmount]
      properties:
        categoryId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        rentalPrice:
          type: number
          format: double
        rentalPeriod:
          $ref: '#/components/schemas/RentalPeriod'
        depositAmount:
          type: number
          format: double
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...